+++ examples/complete/locals.tf
@@ -1,7 +1,20 @@
 locals {
   tags            = merge({ "Name" = var.name }, var.tags)
-  azs             = data.aws_availability_zones.available.names
   service_account = data.aws_elb_service_account.main.arn
-  public_subnets  = [cidrsubnet(var.cidr_block, 8, 1), cidrsubnet(var.cidr_block, 8, 2), cidrsubnet(var.cidr_block, 8, 3)]
-  private_subnets = [cidrsubnet(var.cidr_block, 8, 4), cidrsubnet(var.cidr_block, 8, 5), cidrsubnet(var.cidr_block, 8, 6)]
+  public_subnets  = flatten(local.public_subnet_id)
+  private_subnets = flatten(local.private_subnet_id)
+  vpc_id          = data.aws_vpc.supporting.id
+  private_sub_azs = flatten(local.subnet_azs)
+
+  public_subnet_id = [
+    for i in data.aws_subnet.public : i.id
+  ]
+
+  private_subnet_id = [
+    for i in data.aws_subnet.private : i.id
+  ]
+
+  subnet_azs = [
+    for i in data.aws_subnet.private : i.availability_zone
+  ]
 }